refactor(ShowSelector): break 491-line component into list parts
- ShowListItem (with inline delete confirm), TVDBResultButton, EpisodeListItem
  extracted from the inline .map() blocks
- NewEpisodeForm owns its own form state + create logic (was inline in the
  component)
- Orchestrator keeps the existing useShowSelector hook
- Folder with index.tsx; SubtitleUploader import unchanged
- Preserves this component's plain-Tailwind styling (not yet on the CENA theme)

Verified: tsc clean, lint 0 errors, build succeeds. No behavior change.
